move most of implementations into InMemoryTableScanExec
